backend/requirements.txt
# FastAPI Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Email
aiosmtplib==3.0.1

# PDF Generation
reportlab==4.0.9
arabic-reshaper==3.0.0
python-bidi==0.4.2

# AI Evaluation (OpenAI)
openai==1.12.0
httpx==0.26.0

# File Upload
python-multipart==0.0.6
aiofiles==23.2.1

# Utilities
python-dotenv==1.0.0
